requirements.txt
uvicorn
fastapi
twilio
openai
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
openai-whisper